Working stats and items!!!!
That was fast. There was one major bug but I found and fixed it. Everything is hooked up so items can be picked up, the stats manager will keep the stats up to date based on what items you're holding, and the rigidbody reads the tank's stats to adjust its behavior. And each of those is its own system so nothing is too interconnected! I think I might do something to make stats a res tho tbd