Cleanup (this one has summary)
There is still more to do. Mostly with fine tuning of what to make signals in the game_loop

Essentially the last three or four pushes was a HUGE reorganization of the top level classes of the game. I made like 11 autoloads to handle all the top level game interactions. This makes it really easy to call up and make things happen. The main point of this was organization and to do this I made my goal to turn GameManager from a "doer" into a "delegater". Now, GameManager has central loop that manages the flow of the game by calling a few functions and awaiting signals. Those functions call the other autoloads to do things. This makes everything more readable and understandable. These changes also cascaded good organization into lower aspects of the game, several changes were made to clean up tank (most notably the addition of the TankProfile/PlayerProfile system). Also a few changes were made to stats handling (created BoostFuncLibrary). There are still a few more TODOs in GameManager but the bulk of the transition is complete.